requirements.txt
pycryptodome==3.9.0 
cryptography 
paramiko
